SkillHub/skills/whisper-transcription/scripts/requirements.txt
openai-whisper>=20231117
torch>=2.0.0
ffmpeg-python>=0.2.0
click>=8.0.0
